Remove wine (for x86_64 size issue), dosbox and wget
By dropping wine I keep a bunch of extra libs from being included in the
x86_64 spin so that it should be very close to the size of the i686 spin.

dosbox is an emulator and I think we should be showing off Fedora games
rather than 3rd party games.

wget is not needed now that the spin is based off the desktop. Some way to
grab online content will be included.