Remove games that are only playable with 3rd party downloads.
bolzplatz2006, clonekeen, rott-shareware and vavoom are only playable with
downloads which doesn't work very well with a live DVD.
quake3 is playable with openarena, but currently that isn't included because
of size. If that changes, including openarena will bring it in without
a specific mention, so it should still come off the list.