yum in post of livecd-desktop removed.
Packages should not be removed in post. For the nightly composes the yum remove
is failing in any case. If it were to start working, it could potentially
break several spins that include the livecd-desktop as a base ks.
The syslinux dependency is expected to be fixed the GA release, so this
isn't going to be needed to keep perl out of the livecd-desktop spin.
Oversized CD images can still easily have their functionallity tested.