qa-testday: Add favorites to the same override file desktop uses.
It looks like the override for favorites working depending on the
override file from qa-testday being processed after that from
live-desktop. It worked in testing, but might not work later.
So instead use append and the same name. This will put the qa-testday
definition after the live-desktop one in the same file.