qa-testday: %post sections are not run in the order I expected.
Both qa-testday.ks and fedora-live-desktop.ks were trying to modify
/usr/share/glib-2.0/schemas/org.gnome.shell.qa-testday.gschema.override
by appending to the end of it. However, even though fedora-live-desktop.ks
was %included at the start of qa-testday.ks, the %post section from
fedora-live-desktop.ks was run second, so that its favorite-apps setting
was taking prference.

So I am going back to including a vender string in the name of the
override file. This seems to result in the config from qa-testday.ks
taking precedence. But I don't know if this is based on alphabetic
ordering by filename or just random happenstance. I have a question
pending on the kickstart list to see if I can get more information
on %post ordering.