qa-testday: Resolve overriding of favorite-apps issue
favorite-apps was set in persistant config for qa-testday, but only for
the live enviroment in live-desktop. The latter happens later since it
happens at run time, not build time. So a new section was added to
do a runtime override as well so the the qa-testday favorites are used
both in the live environment and after a test install.