[games] Don't include libreoffice in the games spin
libreoffice was added to the livecd desktop spin very late in the
development cycle and there isn't a lot of time to futz around with
this as we are freezing today. The change would likely put the games
spin oversize. For now just don't include libreoffice stuff. Later
we may want to keep it and adjust the included games.