Fix GDM auto-login for all GNOME-based spins
Using >> means we have a custom.conf file that looks like this:

[daemon]
...
[daemon]
AutomaticLoginEnable=True
AutomaticLogin=liveuser

Which is the first reason why it's failing. Change the cat append to replace as
the other sections are optional.

Additionally, the livesys legacy service is not completing before gdm is being
started, which means the config file is being read before it has been altered.

This can be fixed by adding a LSB-style "X-Start-Before: display-manager" in the
header of livesys.