Move forced excludes out of repo commands
Because pungi is an inclusive depsolver we can exclude some packages it
brings in and still get all of what we need. In particular we need to do
this to save space.

At one time only excludes on the repo commands were truly forced. It
had been the case that packages excluded in the %packages section
could still be included to satisfy dependencies. This is no longer the
case.

Doing the excludes in the packages sections allows us to list the
packages just once and it makes overriding the repo commands simpler
since the excludes don't need to be repeated.