Switch from using auto tools to just building a new tarball
The previous system required making commits in order to do a new
build. This caused problems because we were using the git repo
directly for fedora builds and we couldn't use that exact version
when building matching packages because of needing to make new
commits in order to do the build.

Having the spec file in here also was unnecessary. It is easy to
just maintain the spec file in the Fedora package.

I still want to do a bit more future work with this to pick up
the version from a git tag.