Fix up repo definitions
Someone noticed that I messed up the repo definitions for live images,
but the fix wasn't right for the live images. The install image uses
a source repo, but the live images don't. The repo commands for the
install image are in the install image itself. It is also safe to
include the updates repo since an updates repo is created for branched
that has no packages and stays that way until release is gold. That
way the same repo set can be used before and after release. Also a
commented out repo definition for updates-testing is provided to help
people who want to enable that.