fedora-live-base.ks: do not hardcode fedora-release in %post
Do not hardcode the fedora-release package name in the %post scriptlet
(in the place that imports the Fedora RPM GPG key). Instead, use
"--whatprovides system-release", which is strictly equivalent if
fedora-release is used, but also works when the *-release package is
named differently.

This makes live easier for remixes, which have to replace fedora-release
with generic-release or ${remixname}-release. They would otherwise have
to edit fedora-live-base.ks.