grow base partition for cloud and vagrant images
We control the actual size of the virtual disks with options on the
koji command line.  This change will allow the Vagrant root
partition to grow to the 40 GB we allocate in the koji image build
while the base cloud image will remain essentially unchanged, as it
is set to 3 GB in the rel-eng koji call.