fedora-live-kde-base.ks: Don't shipped cached PackageKit metadata.
This effectively reverts the recent change by rdieter, without undoing
the refactoring.

As per the IRC discussion, it looks like caching the metadata is not all
that helpful with Apper or Muon (and I doubt it is actually helpful with
ANY frontend, because updates will necessarily be outdated, and even the
Everything repo usually changes one last time after the last RC, to
officially push packages that the RC took from a side repo), we would
only be increasing our spin size with stale metadata.