Make sure chronyd runs with the correct config
We need to have chronyd start after livesys has finished so that
the config for chronyd gets rewritten before it starts. If not it
will overwrite the system clock with a time that will be incorrect
(US eastern stored as local time instead of UTC) for most people.
This fixes bug 1018162.