move arm boot partition to ext3
With e2fsprogs after 1.43 the 64bit and metadata_csum features are
enabled by default. These features are not currently supported in
u-boot and the 64bit feature introduces changes such that it cannot
be read by implementations that do not support it. U-Boot does not
support the functionality and hence now won't mount it just in case
it corrupts the filesystem, which is a reasonable response, this how
ever stops us from booting when we have a ext4 /boot file system
which means basically we end up with a pot plant. Go back to using
ext3 for the time being as the mkfs.ext3 option doesn't enable these
features and we get booting systems!! YAY \o/